requirements.txt
bidict==0.22.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.2
Flask-SocketIO==5.3.2
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
Pillow==9.4.0
python-dotenv==0.21.1
python-engineio==4.3.4
python-socketio==5.7.2
requests==2.28.2
urllib3==1.26.14
websocket-client==1.5.1
Werkzeug==2.2.2
zope.event==4.6
zope.interface==5.5.2
